requirements.txt
tensorflow>=2.15
scikit-learn>=1.4
umap-learn>=0.5
hdbscan>=0.8
rembg>=2.0         # background removal
matplotlib>=3.8
jupyterlab>=4.1
nbstripout>=0.7
pytest>=8.0
onnxruntime>=1.17
